Renaming typing to typeutils
