+++ drivers/base/driver.go
@@ -3,7 +3,7 @@ package base
 import (
 	"github.com/piyushsingariya/shift/protocol"
 	"github.com/piyushsingariya/shift/types"
-	"github.com/piyushsingariya/shift/typing"
+	"github.com/piyushsingariya/shift/typeutils"
 )
 
 type Driver struct {
@@ -30,7 +30,7 @@ func (d *Driver) UpdateState(stream protocol.Stream, data types.RecordData) erro
 	if cursorVal, found := data[stream.Cursor()]; found && cursorVal != nil {
 		// compare with current state
 		if stream.GetState() != nil {
-			state, err := typing.MaximumOnDataType(datatype, stream.GetState(), cursorVal)
+			state, err := typeutils.MaximumOnDataType(datatype, stream.GetState(), cursorVal)
 			if err != nil {
 				return err
 			}
