+++ protocol/interface.go
@@ -41,7 +41,7 @@ type BulkDriver interface {
 
 // JDBC Driver
 type JDBCDriver interface {
-	FullLoad(stream protocol.Stream, channel chan<- types.Record) error
+	FullLoad(stream Stream, channel chan<- types.Record) error
 	GroupRead(channel chan<- types.Record, streams ...Stream) error
 	SetupGlobalState(state *types.State) error
 	StateType() types.StateType
