+++ types/data_types.go
@@ -1,12 +1,6 @@
 package types
 
-import "github.com/apache/arrow/go/v16/arrow"
-
 type DataType string
-
-func (typ DataType) ToArrow() arrow.DataType {
-	return dataTypeToArrow(typ)
-}
 
 const (
 	NULL      DataType = "null"
@@ -21,26 +15,3 @@ const (
 )
 
 type RecordData = map[string]any
-
-func dataTypeToArrow(typ DataType) arrow.DataType {
-	switch typ {
-	case NULL:
-		return arrow.Null
-	case INT64:
-		return arrow.PrimitiveTypes.Int64
-	case FLOAT64:
-		return arrow.PrimitiveTypes.Float64
-	case STRING:
-		return arrow.BinaryTypes.String
-	case BOOL:
-		return arrow.FixedWidthTypes.Boolean
-	case OBJECT:
-		return &arrow.MapType{}
-	case ARRAY:
-		return &arrow.ListType{}
-	case TIMESTAMP:
-		return arrow.FixedWidthTypes.Timestamp_ns
-	default:
-		return arrow.BinaryTypes.String
-	}
-}
